requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
banks==2.1.2
blinker==1.9.0
bm25s==0.2.12
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
dj-database-url==2.3.0
Django==4.2.11
django-appconf==1.1.0
django-cors-headers==4.7.0
django-cryptography==1.1
django-extensions==4.1
django-fernet-fields==0.6
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.1
griffe==1.7.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
llama-index-core==0.12.33.post1
llama-index-retrievers-bm25==0.5.2
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.0
